requirements.txt
pywebview==4.4.1
pywin32==306
pyyaml==6.0.1
requests==2.31.0
psutil==5.9.5
packaging==23.1
